Add ENABLE_GUI option to control GUI builds and dependency checks
Invoking cmake with -DENABLE_GUI=ON will require QT/QJSON.  Setting it
to OFF will never build the GUI, even if QT and QJSON is available.  The
default setting is AUTO, whereby QT/QJSON will be used if available (ie.
the current behavior).

Inspired on Tim Harder's patch for Gentoo and the information on
http://www.gentoo.org/proj/en/qa/automagic.xmltest , but modified to not
change the current default automagic behavior.