tracedump: Add tri-state --color option (auto, always, or never)
This follows a pattern similar to that provided by other tools, (such
as grep). The default auto mode colorizes if stdout is a tty. The always
and never modes explicit enable or disable colorizing.

The old --no-color and --no-colour options are deprecated, (no longer
documented but still supported to avoid any regressions of scripts).

Signed-off-by: José Fonseca <jose.r.fonseca@gmail.com>