egl: Merge various enum types
Move most enums to EGLenum and add EGLattrib as an alias of EGLenum.  The idea
is that the same enum value may be used (or abused) in several places, and we
were forced to list the same enum in multiple types.

The difference between EGLenum and EGLattrib is that EGLattrib maps to EGLint
instead of EGLenum in C.