Display captured snapshots as thumbnails.
Once snapshots are captured after being generated by glretrace, they are
then displayed in the GUI app, qapitrace.

The retracer passes a list of thumbnails to the main window. The main
window in turn binds those thumbnails to the ApiTraceFrames for later
display.

When the ApiTrace events are displayed, if a frame is being display, its
thumbnail is queried. That thumbnail is then displayed at the
front of the line for a frame.