Add trim support to qapitrace GUI app.
This patch adds support for trimming of traces via the qapitrace GUI.
We enhance the GUI by adding a Trim entry to the Trace menu. When the
user selects either a frame or a call, the "apitrace trim" command
will be invoked to trim all calls after the selected call or frame.
New trace files are created automatically and named according to the
last call number in the trimmed trace.

For example, if the original trace file is:
   /path/foo.trace
and the trace is trimmed after call 1234, the name of the trimmed trace
file will be named:
   /path/foo.1234.trim.trace

Upon trimming, the trimmed trace file will be loaded into qapitrace.

First, we enhance the Trace menu of the GUI app.

Next, we add a TrimProcess class. This is modelled on the TraceProcess
class, but takes into account differences and simplifications.

Next, we tie the TrimProcess class into the main window by accessing
that class and its members appropriately as well as tieing in message
communication with that class.

Finally, we add a reference to the source of TrimProcess to the make
system.

Signed-off-by: José Fonseca <jose.r.fonseca@gmail.com>