Don't show images where there is no difference.
This makes for a much more compact view for the common case where
image mismatches are rare, and makes the mismatches stand out more
making them easier to notice.

To avoid hiding information irretrievably, the filename is made into a
link to the image so that it can be easily seen if desired.

Signed-off-by: José Fonseca <jose.r.fonseca@gmail.com>