apitrace diff-images: Print one line for each file being compared.
This is important so that the test suite can verify that "apitrace
diff-images" is actually doing something. The tricky case is when no
images are found to compare. In this case, diff-images exits with
value 0, (no differences), which is valid, but the test suite wants to
know that no comparison happened since that's unexpected, (and should
flag a failure in the test suite).