gui: Notify retracer of missing thumbnails to capture.
Now that the trace knows what thumbnails are needed, that information
needs to be communicated to the retracer to actually capture them.

The trace is modified to iterate through the missing thumbnails. For
each of these thumbnail indices, the trace calls a callback to notify
it of the thumbnail.

I don't really like my solution to this but I couldn't figure out a
better one. What I wanted to do was have the trace hand out an iterator
along the lines of what Java or C# do, but I couldn't figure out how
to create one. If the reviewers of this code have a suggestion, I'm all
ears.

The main window provides the callback and acts as an intermediate to the
retracer. The callback receives the missing index and passes it to the
retracer, which then adds the index to the collection of missing thumbnail
indices it is gathering.