gui: Prepare to receive sparse thumbnail data.
The previous implementation assumed that all frame thumbnails would be
provided to the trace by the retracer. Now, not only are just a subset of
frame thumbnails being provided, but call thumbnails may also be present,
depending on where the user was looking.

Therefore, rather than process a list of thumbnails, one for each frame,
this patch modifies the code to process a hash table of thumbnails which
is addressed by the call index.

In addition, the code now needs to extract the call index that was encoded
by glretrace as an image comment. Therefore, the PNM parser will now
attempt to interpret the comment in the PNM header as a number and return
it in addition to the other header data.

In order for Qt to manage the modified message data, main.cpp now has to
define a different data packet as well (ImageHash). I tried to not define
that new type but the Qt preprocessor didn't like the two type parameters
I was using.

Finally, ApiTrace::bindThumbnails is made more generic to deal with the
sparse data in the hash table as well as potentially assign call
thumbnails as well as frame thumbnails.

v2: Attach inner thumbnails to the right call.  Calls may be filtered,
so it's not possible to find the right call merely by indexing.  -- Jose