Avoid walking off array if given a negative value for array size.
Obviously, a negative size is bogus, but we want apitrace to be robust
enough to capture even bogus usage of the API, (so that the user can
investigate and fix the bogosity in the application).

In this case, if the user specifies a negative size for an array,
simply don't write anything out for the array content.

Signed-off-by: José Fonseca <jose.r.fonseca@gmail.com>