Recreate egl surface when glViewport for watching framebuffer in qapitrace's surface tab.
It is quirk of mesa egl 1.4 with Nvidia-dir. It is ok with intel-dri.

After resizing XWindow, mesa egl does not know new size of the surface.
If you call eglQuerySurface(currentDisplay, currentSurface, EGL_WIDTH, width),
width is the width at the time to create the egl surface, not current width.
However, glx can know what the size of window well.