trim: Add a new --print-callset option.
This is useful whenever one has a need to manually tweak the trim operation.

Simply run:

	apitrace trim --print-callset --frames=<whatever> --calls=<whatever>

as desired. Then take the resulting callset that is printed and run:

	apitrace trim --calls=<callset> --exact

This should yield an identical trimming operation. Then this operation
can be tweaked by appending call ranges to the --calls option,
deleting rangess from the --calls option, or by appending a new
--frames option.