Use ndk-android-r7 for FirefoxOS builds.
Commit 7445733 added support for FirefoxOS, but appears to have been using
a newer version of the NDK than we typically build with.  The codeaurora.org
prebuilts repo we are tracking has not pulled in r8 yet and only provides r7.

For the time being it would be helpful to drop back to r7.  Admittedly, we
should probably have a better mechanism for selecting the latest version
available.