egltrace/android: Fix tracing Zygote processes (v2)
Fixes egltrace.so on Android 4.2 x86.

Do not call trimDirectory() on the proc_name of Zygote processes, because
a Zyogote process name never contains a path separator. The proc_name of
a Zygote process is the application's package name (such as
com.exampe.myapp) because ActivityManager rewrites argv[0].

There exists an undiagnosed problem with trimDirectory, but I have been
unsuccessful diagnosing it. On the other hand, the call to trimDirectory
isn't needed and its removal fixes the bug's symptoms.

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>