gltrace: Expose marker functions when tracing is disabled.
Matches the output of change proposed by Peter Lohrmann in issue #138,
but with slightly less new code.

This is achieved by adding a new hook point, doInvokeFunction (could not
think of a better name), used to generate the call to the real function,
both when trace is enabled or disabled.