apitrace: Allow multiple --calls and/or --frames input options
Input for multiple files or ranges that seemed intuitive did not work

Only initialize CallSet list on first option call. If the option is
repeatedly used, continue adding to the same list.
e.g., apitrace trim --calls=@foo.calls --calls=0-20 --calls=...

The "CallSet::CallSet(const char *string)" constructor determined if the
option was a file or a range and then made the relative call to populate
the CallSet list. Rename that constructor to a public member function
"CallSet::merge(const char *string)" and call it from the command line
parser istead of constructing a new object each time the option is used.
In the renamed "merge" function, add a check to create the new CallSet
object only when the option is called the first time.

Modified modules that use this option: trim, dump, pickle, retrace

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>