qapitrace: add Apply button to settings form
It's less annoying to have an apply button for seeing the affect
of the the different filters in the settings form rather than
bringing up the form over and over again.

TESTING
I tested this in a branch where I'm doing some other qapitrace
development and everything worked as expected. But after moving
the (Apply-button-specific) changes to a separate branch to push,
I noticed the filter for
"Only show the following events"
for
"glDraw|glVertex|glBegin|glEnd"

wasn't working correctly (not showing all the selected events)
I tested against the current upstream master and it doesn't work
in the same way. I tried to go back to see if it ever worked
correctly but ran into other problems.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>