qapitrace: Highlight color too dark to read row text
PROBLEM
On my debian system, the default highlight color is dark blue against
black text and is virtually unreadable

% uname -a
Linux gok 3.13-1-amd64 #1 SMP Debian 3.13.5-1 (2014-03-04) x86_64 GNU/Linux

SOLUTION
On initialization, set treeview highlight color to a lighter shade.

COMMENTS
This could be extended as a command-line option or added to the Options
dialog window to allow the user to select highlight a color. A gradient
colored highlight gives a nice look also (image attached).

CAVEAT
The highlight color may be ignored on debian systems that haven't been
recently updated. I do a weekly software update on my system and a few
weeks ago when I tried changing the hightlight color it was ignored. Now
it works so it may be ignored on systems that havent' been updated
recently.
The reason may be because Qt/GTK don't (didn't) use Xrdb as explained
here
http://stackoverflow.com/questions/16254895/x11-how-to-get-system-colors#tab-top

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>