glretrace: Create decorator parser class to allow variety of parsers.
The parser class is abstracted so that various looping parsers  can be
added as decorators.  Add the last frame looping parser as the first
looping parser.  Fix the singlethreaded case so it functions with the
looping parser just like the multithreaded case.

v2: Simplify/cleanup -- Jose Fonseca <jfonseca@vmware.com>