glretrace: Fix `unavailable function eglGetCurrentContext` with Catalyst drivers.
AMD's Catalyst driver's export eglGetCurrentContext symbol in fglrx_dri.so

getDrawableBounds was calling

  dlsym(RTLD_DEFAULT, "eglGetCurrentContext")

and detecting the symbol as present, while glproc was calling

  dlsym(dlopen("libGL.so.1",...), "eglGetCurrentContext")

and aborting because it could not find it.

This change prevents this mismatch, and should therefore fix issue #198.

Nevertheless, the approach used in glstate to detect the presence/absence
of EGL still seems a tad brittle...